+++ changelog.md
@@ -1,6 +1,7 @@
 ### 2.0
 - More efficient temporary file handling
 - Added Undo & Redo
+- Added Popup-Menus to the Effects
 - Added Vertical and Horizontal Flip
 - Added 180-degree rotation
 - Removed Distortion Flip Effect & Flip Effect